+++ sharepoint-adfs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Enhancements & bug-fixes - Published in February 9, 2021
 
-* Implement temporary workaround to blocking cChoco issue https://github.com/chocolatey/cChoco/issues/151
+* Update DSC module cChoco from 2.4 to 2.5 to fix issue <https://github.com/chocolatey/cChoco/issues/151>
 
 ## Enhancements & bug-fixes - Published in December 10, 2020